ignoreCli option.. fixed tests.. removed .listen() and close().. refactor
